+++ examples/get_all.rs
@@ -1,4 +1,5 @@
 fn main() {
+    // Retrieve all listeners
     if let Ok(listeners) = listeners::get_all() {
         for l in listeners {
             println!("{l}");